fix: prefer u64 instead of usize
In the current code base `usize` is used a lot, but most of the time it is
use as if it were a u64. Hence use u64 for new code.

In the future the code should be checked if u64 sizes are always needed, or
if sometimes smaller sizes would be sufficient.

+++ filecoin-proofs/src/api/seal.rs
@@ -364,7 +364,7 @@ pub fn verify_seal(
         &public_inputs,
         &proof,
         &ChallengeRequirements {
-            minimum_challenges: POREP_WINDOW_MINIMUM_CHALLENGES.load(Ordering::Relaxed), // TODO: what do we want here?
+            minimum_challenges: POREP_WINDOW_MINIMUM_CHALLENGES.load(Ordering::Relaxed) as usize, // TODO: what do we want here?
         },
     )
     .map_err(Into::into)
